+++ src/front-end/js/components/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Menu from './menu/menu.jsx';
-import MyMap from './map/map.jsx';
-import Footer from './footer/footer.jsx';
+import Menu from './menu.jsx';
+import MyMap from './map.jsx';
+import Footer from './footer.jsx';
 
 class Harley extends React.Component {
     render() {
