+++ src/front-end/less/components.less
@@ -1,3 +1,3 @@
-@import "../components/menu/less/menu";
-@import "../components/map/less/map";
-@import "../components/footer/less/footer";
+@import "components/menu";
+@import "components/map";
+@import "components/footer";
