+++ src/front-end/less/components/menu.less
@@ -32,8 +32,8 @@
   }
   aside{
     position: absolute;
-    top:0;
-    right:-300px;
+    top:-20px;
+    right:-330px;
     width:300px;
     height:100vh;
     padding:15px;
@@ -41,6 +41,7 @@
     background-color: @background-light;
     -webkit-transition: all 500ms;
     transition: all 500ms;
+    z-index: 9999;
     ul{
       margin:0;
       li{
@@ -61,7 +62,7 @@
       }
     }
     &.opened{
-      right:0;
+      right:-30px;
     }
   }
 }
