+++ src/front-end/less/reset.less
@@ -1,0 +1,4 @@
+.row.fluid{
+  width:100%;
+  max-width:100%;
+}
